# docs/swagger.yaml
basePath: /api
definitions:
  main.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  main.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  main.Product:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
  main.SuccessResponse:
    properties:
      message:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API sederhana untuk manajemen kasir (Produk & Kategori).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kasir API
  version: "1.0"
paths:
  /category:
    get:
      consumes:
      - application/json
      description: Mengembalikan list array dari semua kategori yang tersedia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Category'
            type: array
      summary: Mendapatkan semua kategori
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Menambahkan data kategori baru ke database
      parameters:
      - description: Data kategori
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/main.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Membuat kategori baru
      tags:
      - categories
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Menghapus kategori
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Mengambil detail satu kategori
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Mendapatkan kategori berdasarkan ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Mengupdate data kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data kategori
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/main.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Mengupdate kategori
      tags:
      - categories
  /health:
    get:
      consumes:
      - application/json
      description: Mengecek status API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - health
  /product:
    get:
      consumes:
      - application/json
      description: Mengembalikan list array dari semua produk yang tersedia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Product'
            type: array
      summary: Mendapatkan semua produk
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Menambahkan data produk baru ke database
      parameters:
      - description: Data Produk
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Membuat produk baru
      tags:
      - products
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Menghapus produk
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Mengambil detail satu produk
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Mendapatkan produk berdasarkan ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Mengupdate data produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data Produk
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Mengupdate produk
      tags:
      - products
swagger: "2.0"